pyproject.toml
[project]
name = "currency-exchange"
version = "0.1.0"
description = "Currency Converter API with transaction tracking"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.41",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "aioresponses>=0.7.5",
    "respx>=0.20.2",
    "SQLAlchemy>=2.0.41",
]
